Bresenham's line algorithm has been implemented
The code is heavily commented to describe the process of the algorithm.
It's been tested with some cases and works as intended. Further testing
will be done to ensure the quality of the algorithm.

Will possibly redact the algorithm process explanation at the beginning
of the function as it is a bulk of information and probably don't need
to be explain in the codebase.