Raster has been refactored.
It now implements the Color and Vector2Int classes to provide a higher
level of abstraction and eliminate lengthy parameters in its functions.
Only the signatures of the methods have been refactored and the actually
implementation of them reflected the parameters. The code itself will
probably be refactored for better readability.