PPMFile still in progress.
Creating a PPMFile checks if the path (if specified) is valid and if a
.ppm file already exists. If it does, it duplicates the name of the
original with an appended 'Copy(n)' to it where n is the next available
number of a filename that doesn't exist. Additionally, it replaces the
extension name if the wrong extension is specified or none at all.